Adpartner & Mediaimpact adapters: share common sync helpers
